% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clark_wright.R
\name{clark_wright}
\alias{clark_wright}
\title{Clark-Wright Algorithm (CVRP solver)}
\usage{
clark_wright(demand, distances, vehicles)
}
\arguments{
\item{demand}{A numeric vector consisting of "demands" of sites. The length
of the vector equals the number of sites \code{N} with demands. The
units of the demand values need to match the units of the
vehicle capacity values.}

\item{distances}{An object of class \code{dist} (created by \code{\link[stats:dist]{stats::dist()}}) with
\code{N + 1} locations, describing the distances between sites.
The zeroth location refers to the source.}

\item{vehicles}{A data frame describing available vehicle types and
their respective capacities. One row per vehicle type.
The data frame is expected to have two columns:
* \code{n} - Number of available vehicles. This can be set to
\code{NA} if the number is "infinite" (i.e. effectively
the maximal integer value on your machine.)
* \code{caps} - The vehicle capacity in same units as \code{demand}.
The order of the data frame is relevant and determines
the prioritization of vehicle assignments to runs (in case
two or more vehicle types are eligible for assignment the
"first" vehicle is chosen).}
}
\value{
Returns a data frame of the same length as \code{demand} with columns
* \code{run_id} - Identifies the run the site is assigned to.
* \code{order}  - Provides the visiting order within each run.
* \code{vehicle_id} - The vehicle index (as provided in \code{vehicles})
associated to the run.
}
\description{
Finds a quasi-optimal solution to the Capacitated Vehicle Routing
Problem (CVRP). It is assumed that all demands can and will be satisfied by a
single source.
}
\examples{
demand <- c(3, 2, 4, 2)

pos <-
  data.frame(
    pos_x = c(0, 1, -1, 2, 3),
    pos_y = c(0, 1, 1, 2, 3)
  )

clark_wright(
  demand,
  dist(pos),
  data.frame(n = NA_integer_, caps = 6)
)
}
