% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvrplib.R
\name{cvrplib_download}
\alias{cvrplib_download}
\title{CVRPLIB problem instance downloader}
\usage{
cvrplib_download(qualifier, cvrplib_url = "https://galgos.inf.puc-rio.br")
}
\arguments{
\item{qualifier}{The qualifier of the problem instance. E.g. "tai150d".
This can either be inferred directly from the website or by the output of
\code{\link[=cvrplib_ls]{cvrplib_ls()}}.}

\item{cvrplib_url}{The base URL of the CVRPLIB website. Defaults to https://galgos.inf.puc-rio.br.}
}
\value{
Returns a "\code{cvrplib_instance}" object which contains CVRPLIB problem
instance data.
}
\description{
\href{https://galgos.inf.puc-rio.br/}{CVRLIB} offers a selection of
CVRP problem instances. This function downloads the instance data and
conveniently makes it available to be fed into solver functions, e.g. with
\code{\link[=clarke_wright_cvrplib]{clarke_wright_cvrplib()}}. The primary purpose for those instances is
benchmarking / comparing speed as well as performance of solvers.
}
\seealso{
Other cvrplib: 
\code{\link{clarke_wright_cvrplib}()},
\code{\link{cvrplib_ls}()}
}
\concept{cvrplib}
