% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cvrplib_Tai}
\alias{cvrplib_Tai}
\title{CVRPLIB - CVRP instance data}
\format{
\subsection{\code{cvrplib_Tai}}{

A list of CVRP instances as "\code{cvrplib_instance}" objects. The instances
can be directly fed into solver algorithm, e.g. via \code{\link[=clarke_wright_cvrplib]{clarke_wright_cvrplib()}}.
}
}
\source{
\url{http://vrp.atd-lab.inf.puc-rio.br}
}
\usage{
cvrplib_Tai
}
\description{
A collection of CVRP instances by Rochat and Taillard, 1995.
}
\keyword{datasets}
