% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{plot.heumilkr_solution}
\alias{plot.heumilkr_solution}
\title{Plot a CVRP solution}
\usage{
\method{plot}{heumilkr_solution}(x, ...)
}
\arguments{
\item{x}{A "\code{heumilkr_solution}" object, typically obtained by \code{\link[=clarke_wright]{clarke_wright()}}.}

\item{...}{Not used.}
}
\description{
Plot a CVRP solution
}
