% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarke_wright.R
\name{clarke_wright}
\alias{clarke_wright}
\title{Clark-Wright Algorithm (CVRP solver)}
\usage{
clarke_wright(demand, distances, vehicles)
}
\arguments{
\item{demand}{A numeric vector consisting of "demands" indexed by sites.
The \code{i}th entry refers to the demand of site \code{i} (and the length
of the vector equals the number of sites \code{N} with demands). The
units of the demand values need to match the units of the
vehicle capacity values.}

\item{distances}{An object of class \code{dist}, created by \code{\link[stats:dist]{stats::dist()}}, with
\code{(N + 1)} locations describing the distances between individual
sites. The first index refers to the source site. The \code{(i+1)}'th
index refers to site \code{i} (as defined by \code{demand}).}

\item{vehicles}{A data frame describing available vehicle types and their respective
capacities. One row per vehicle type. The data frame is expected to have
two columns:
\itemize{
\item \code{n} - Number of available vehicles. This can be set to \code{NA} if the
number is "infinite" (i.e. effectively the maximal integer value
on your machine.).
It is recommended to keep at least one vehicle type as "infinite",
otherwise the solver might raise a run time error due to initially
not having enough vehicles available (even though the final
solution might satisfy the availability restrictions).
\item \code{caps} - The vehicle capacity in same units as \code{demand}.
}

The order of the data frame is relevant and determines the prioritization
of vehicle assignments to runs (in case two or more vehicle types are
eligible for assignment the "first" vehicle is chosen). In a typical scenario
"more expensive" vehicles should be further down in the list (so the cheaper
one is chosen in case there is doubt). Since higher capacity vehicles
usually involve higher costs sorting the data frame by capacity is usually
a good rule of thumb.}
}
\value{
Returns a data frame with one row per site-run combination.
\itemize{
\item \code{site} - The site index (as provided in \code{demand}) associated
to the run.
\item \code{run} - Identifies the run the site is assigned to.
\item \code{order}  - Integer values providing the visiting order within each run.
\item \code{vehicle} - The vehicle index (as provided in \code{vehicles}) associated
to the run.
\item \code{load} - The actual load in units of \code{demand} on the particular run.
}

Unless a site demand exceeds the vehicle capacities it is always assigned
to only a single run.
}
\description{
Finds a quasi-optimal solution to the Capacitated Vehicle Routing
Problem (CVRP). It is assumed that all demands will be satisfied by a
single source.
}
\details{
See the original paper,
\href{http://dx.doi.org/10.1287/opre.12.4.568}{Clarke, G. and Wright, J.R. (1964)},
for details.
}
\examples{
demand <- c(3, 2, 4, 2)

pos <-
  data.frame(
    pos_x = c(0, 1, -1, 2, 3),
    pos_y = c(0, 1, 1, 2, 3)
  )

clarke_wright(
  demand,
  dist(pos),
  data.frame(n = NA_integer_, caps = 6)
)
}
