% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cvrplib_F}
\alias{cvrplib_F}
\title{CVRP instance data by Fisher, 1994}
\format{
\subsection{\code{cvrplib_F}}{

A list of CVRP instances as "\code{cvrplib_instance}" objects. The instances
can be directly fed into solver algorithm, e.g. via \code{\link[=clarke_wright_cvrplib]{clarke_wright_cvrplib()}}.
}
}
\source{
\url{http://vrp.atd-lab.inf.puc-rio.br}
}
\usage{
cvrplib_F
}
\description{
A collection of CVRP instances by Fisher, 1994, provided courtesy of
\href{http://vrp.atd-lab.inf.puc-rio.br}{CVRPLIB}. See
\href{http://vrp.atd-lab.inf.puc-rio.br}{CVRPLIB} for visualizations of the
instances and their solutions as well as a multitude of alternative instance
data.
}
\keyword{datasets}
