% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvrplib.R
\name{cvrplib_ls}
\alias{cvrplib_ls}
\title{List available CVRPLIB online data}
\usage{
cvrplib_ls()
}
\value{
A vector of data set qualifiers which can be used with \code{\link[=cvrplib_download]{cvrplib_download()}}.
}
\description{
Scrapes the CVRPLIB website to look for available data sets.
This function call can take some time.
}
\seealso{
Other cvrplib: 
\code{\link{clarke_wright_cvrplib}()},
\code{\link{cvrplib_download}()}
}
\concept{cvrplib}
